fix setting TZ in tests
This was regressed in 139beba which set `PYTHONUNBUFFERED` to fix other
test issues. When setting multiple environment variables in this way
they need to be delimited with semicolons rather than spaces. As it is
`PYTHONUNBUFFERED` is being set to `1 TZ=America/Chicago`. The CMake docs
are not as clear about this as they probably should be.

This can be verified by throwing together a CTestTestfile.cmake:

    add_test(incorrect_env "printenv" "PYTHONUNBUFFERED" "TZ")
    set_tests_properties(incorrect_env PROPERTIES  ENVIRONMENT "PYTHONUNBUFFERED=1 TZ=America/Chicago")
    add_test(correct_env "printenv" "PYTHONUNBUFFERED" "TZ")
    set_tests_properties(correct_env PROPERTIES  ENVIRONMENT "PYTHONUNBUFFERED=1;TZ=America/Chicago")

When run with `ctest -V`:

    1: Test command: /usr/bin/printenv "PYTHONUNBUFFERED" "TZ"
    1: Environment variables:
    1:  PYTHONUNBUFFERED=1 TZ=America/Chicago
    1: Test timeout computed to be: 10000000
    1: 1 TZ=America/Chicago
    1/2 Test #1: incorrect_env ....................***Failed    0.00 sec
    test 2
        Start 2: correct_env

    2: Test command: /usr/bin/printenv "PYTHONUNBUFFERED" "TZ"
    2: Environment variables:
    2:  PYTHONUNBUFFERED=1
    2:  TZ=America/Chicago
    2: Test timeout computed to be: 10000000
    2: 1
    2: America/Chicago
    2/2 Test #2: correct_env ......................   Passed    0.00 sec