use built-in cmake precompiled header support when available
Ledger supports precompiling headers as a build speed optimization.
This is provided via a custom add_pch_rule macro. Similar functionality
is now built-in to CMake starting with the 3.16 release in November
2019. Let's use this when available to fix #1774 and start towards not
needing to maintain our own implementation.

I originally considered removing the macro fallback but in my tests it
saves enough build time that I think it is worth keeping for now.

Fixes #1774