collapse_posts: Fix some postings getting lost when collapsing
Because the only_predicate and display_predicate are applied to the
original posts, not the collapsed ones, they sometimes yield wrong
results leading to entire transactions getting dropped erroneously.

As this code is only relevant to depth == 0 collapsing (that is, the old
behavior), only follow the displayed_count == 1 code-path when the depth
is 0.