Document how one might pass multiple debug flags
Issue #1062 mentions how someone might pass multiple debug flags to ledger using a regular expression, though it was never mentioned in the documentation from what I could see. I've added a small patch to the relevant section of the documentation with a single example to make this functionality easier to find.