Fix IRC bullet point in README.md
Currently the IRC link is broken in the rendered readme page on Github.
Speculating that Github's markdown renderer might not display the text
as a link since it starts with "irc://" instead of "http://" or similar.

Added some text to the bullet point specifying channel and server.

I left the link in-place in case the above issue is resolved at some
point, or if other websites render this README.md file without the
issue.