Fix =regex note query syntax
The manpage documents `=regex` as equivalent to the `note regex` query
syntax, but the former does not actually work as the parser only handles
an equals sign in the case of `tag type=dining` syntax, and doesn't
handle the case where an equals sign starts a note query.

Fixing this does break queries like `tag type = dining` with spaces
around the equals sign, but that syntax was not intended or documented.

Closes: #2275