Stop making timelogs virtual to avoid balancing
This adds a new flag to the post class(?) to state whether the posting
is for a timelog entry. Then the class' must_balance method is updated
to take that flag in to account. Finally, the timelog file is updated to
create postings with that flag set, allowing them to avoid the balance
requirement.