Always use precompiled headers when enabled
As precompiled headers are a build time optimization, we want to use
them always (if supported), as otherwise we would intentionally slow
down builds due to the big precompiled header that the compiler parses
for every C++ source file.

Also given that precompiled headers are a user configuration option, we
don't want to silently disable it on some condition.