Add an undocumented developer feature: ^ at the beginning of a line
This commit is maintainer prerogative. :) This is an undocumented feature
right now, but I find I use it constantly: If you put a ^ at the beginning of
a line (or on a line by itself), then Ledger will interpret that character as
EOF. I use this when bisecting to find where in a file a lack of balancing
occurs.