Refine epoch handling to fix xact command year directive issue
This change improves how the epoch is used when parsing dates without years.
When the epoch is set to Dec 31 (which happens with year directives), we now
skip the month rollback logic since year directives explicitly set the desired year.

This fixes issue #2413 where xact command wasn't respecting the last year directive.

Note: Tests 1038_1 and 1038_2 still fail due to an interaction between
year directives and 'this month' period parsing. This appears to be a
separate issue that predates this fix and needs further investigation.