feat: improve formatting
1. Move the "print transactions and postings" logic from dispatch into
   journal and posting. This makes it easier to test each, and I tried
   to do so w/o allocating a lot of new or big Strings.

2. Add tests for parsing and formatting journals and postings. See #1.
   This is nice because we can confirm, in one place, that we can parse
   all of the data included in then input, and also that we use that
   data correctly when producing output.

3. Try to improve the "register commodities" step (post parse) to update
   commodities as they are found in the input.

4. Try to use the journals commodities to standardize the formatting in
   the output.

5. Improve formatting of postings to align accounts to 36 cols (or longer
   if needed for some account names) and to right align amounts to 10/12
   cols. Longer amounts are still not handled perfectly, but they're a
   lot closer than they were.

This moves the needle on acceptance tests from 12 passing to 14, and I
suspect that a few more of the "output diffs" have narrowed, too. (ie
we're getting closer, even if we're still not passing)

Notably, 1D275740.test is very close to passing, and may well pass once
default commodity support is added, as well as correctly handling
commodity-aware parsing. Eg, `123,456£` could either be `123,456.00£`
or `123.456£`, depending on how the £ commodity has been defined
previously.