fix: reduce post amount before comparing to assertion
When balance assertions are parsed the "PARSE_NO_REDUCE" flag is not used so the
parsed assertion amount is automatically reduced to the smallest possible unit.

When comparing with the posting the commodity is required to be the same, so if
we post in a higher unit we have to reduce it before comparing with the
assertion.

Fixes #960